Remove ui_old file
